fix: code climate
