+++ app/javascript/shared/mixins/editFolderMixin.js
@@ -1,4 +1,8 @@
 import { mapGetters } from 'vuex';
+import {
+  getInputType,
+  getAttributeInputType,
+} from 'dashboard/helper/customViewsHelper';
 import advancedFilterTypes from 'components/widgets/conversation/advancedFilterItems';
 import languages from 'components/widgets/conversation/advancedFilterItems/languages';
 import countries from 'shared/constants/countries.js';
@@ -39,28 +43,22 @@ export default {
         }))
       );
     },
-    getInputType(key, operator) {
-      if (key === 'created_at' || key === 'last_activity_at')
-        if (operator === 'days_before') return 'plain_text';
-      const type = this.filterTypes.find(filter => filter.attributeKey === key);
-      return type?.inputType;
-    },
     generateValueObjectForFilter(filter) {
       const { attribute_key, filter_operator, values } = filter;
-      const inboxType = this.getInputType(attribute_key, filter_operator);
+      const inboxType = getInputType(
+        attribute_key,
+        filter_operator,
+        this.filterTypes
+      );
       const allCustomAttributes = this.$store.getters[
         'attributes/getAttributesByModel'
       ]('conversation_attribute');
 
       if (inboxType === undefined) {
-        const customAttribute = allCustomAttributes.find(
-          attr => attr.attribute_key === attribute_key
+        const filterInputTypes = getAttributeInputType(
+          attribute_key,
+          allCustomAttributes
         );
-        const { attribute_display_type } = customAttribute;
-        const filterInputTypes = this.generateCustomAttributesInputType(
-          attribute_display_type
-        );
-
         return filterInputTypes === 'string'
           ? values[0].toString()
           : { id: values[0], name: values[0] };
@@ -118,17 +116,6 @@ export default {
         id: values[0],
         name: item ? item[nameKey] : values[0],
       };
-    },
-    generateCustomAttributesInputType(type) {
-      const filterInputTypes = {
-        text: 'string',
-        number: 'string',
-        date: 'string',
-        checkbox: 'multi_select',
-        list: 'multi_select',
-        link: 'string',
-      };
-      return filterInputTypes[type];
     },
   },
 };
