+++ app/javascript/shared/mixins/editSegmentMixin.js
@@ -1,3 +1,7 @@
+import {
+  getInputType,
+  getAttributeInputType,
+} from 'dashboard/helper/customViewsHelper';
 import contactFilterItems from 'dashboard/routes/dashboard/contacts/contactFilterItems';
 import countries from 'shared/constants/countries.js';
 
@@ -25,28 +29,22 @@ export default {
         }))
       );
     },
-    getInputType(key, operator) {
-      if (key === 'created_at' || key === 'last_activity_at')
-        if (operator === 'days_before') return 'plain_text';
-      const type = this.filterTypes.find(filter => filter.attributeKey === key);
-      return type?.inputType;
-    },
     generateValueObjectForFilter(filter) {
       const { attribute_key, filter_operator, values } = filter;
-      const inboxType = this.getInputType(attribute_key, filter_operator);
+      const inboxType = getInputType(
+        attribute_key,
+        filter_operator,
+        this.filterTypes
+      );
       const allCustomAttributes = this.$store.getters[
         'attributes/getAttributesByModel'
       ]('contact_attribute');
 
       if (inboxType === undefined) {
-        const customAttribute = allCustomAttributes.find(
-          attr => attr.attribute_key === attribute_key
+        const filterInputTypes = getAttributeInputType(
+          attribute_key,
+          allCustomAttributes
         );
-        const { attribute_display_type } = customAttribute;
-        const filterInputTypes = this.generateCustomAttributesInputType(
-          attribute_display_type
-        );
-
         return filterInputTypes === 'string'
           ? values[0].toString()
           : { id: values[0], name: values[0] };
@@ -71,24 +69,6 @@ export default {
         default:
           return { id: values[0], name: values[0] };
       }
-    },
-    getValuesName(values, list, idKey, nameKey) {
-      const item = list.find(v => v[idKey] === values[0]);
-      return {
-        id: values[0],
-        name: item ? item[nameKey] : values[0],
-      };
-    },
-    generateCustomAttributesInputType(type) {
-      const filterInputTypes = {
-        text: 'string',
-        number: 'string',
-        date: 'string',
-        checkbox: 'multi_select',
-        list: 'multi_select',
-        link: 'string',
-      };
-      return filterInputTypes[type];
     },
   },
 };
